[Java]: Minor style and warning cleanup.

+++ aeron-tools/src/main/java/uk/co/real_logic/aeron/tools/Stats.java
@@ -134,7 +134,7 @@ public class Stats
             vals = new long[tmpVals.size()];
             for (int i = 0; i < vals.length; i++)
             {
-                vals[i] = tmpVals.get(i).longValue();
+                vals[i] = tmpVals.get(i);
             }
         }
         else if (output instanceof StatsConsoleOutput)
@@ -157,7 +157,7 @@ public class Stats
             vals = new long[tmpVals.size()];
             for (int i = 0; i < vals.length; i++)
             {
-                vals[i] = tmpVals.get(i).longValue();
+                vals[i] = tmpVals.get(i);
             }
         }
         else if (output instanceof StatsVmStatOutput || output instanceof StatsCsvOutput)
